requirements/requirements.txt
amqp==5.3.1
celery==5.5.1
fasttext==0.9.3
kombu==5.5.3
numpy==1.23.5
psycopg2-binary
praw==7.8.1
prometheus_client==0.21.1
python-dotenv==1.1.0
redis>=4.0
requests==2.32.3
sentence-transformers==4.1.0
sqlalchemy